requirements.txt
zenml[server]
apache-airflow==2.7.3
feast
great_expectations
dvc
hydra-core
pytest
pandas
scikit-learn
pendulum
shyaml
mlflow==2.14.1
giskard
evidently
Flask
flask-session2
psycopg2-binary
kaggle
dvc-gdrive
giskard==2.14.0
coverage
gradio==3.50.2
